# openapi.yaml
openapi: 3.0.1
info:
  title: CropURL
  version: 1.0.0

paths:
  /user/login:
    post:
      tags:
        - users
      summary: Logs in the user specified
      operationId: loginUser
      parameters:
      - name: username
        in: query
        description: The user name for login
        required: true
        schema:
          type: string
      - name: password
        in: query
        description: The password for login in clear text
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          headers:
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
        400:
          description: Invalid username/password supplied
          content: {}
          
  /user/logout:
    post:
      tags:
        - users
      summary: Logs out current logged in user session
      operationId: logoutUser
      responses:
        200:
          description: successful logout
          content: {}
  
  /new-link:
    post:
      tags:
        - short_links
      summary: Creates a new pair 'short URL - long URL'
      operationId: NewLink
      requestBody:
        description: Item object that needs to be added to the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Link'
        required: true
      security:
        - api_key: []
      responses:
        200:
          description: successful operation
          content: {}
        400:
          description: Invalid input
          content: {}
      
  
  /links/{shortID}:
    get:
      tags:
        - short_links
      summary: Shows statistics about specific short URL
      operationId: ViewStatistics
      parameters:
      - name: shortID
        in: path
        description: Short URL the user wants to get data about
        required: true
        schema:
          type: string
      security:
        - api_key: []
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Link'
        400:
          description: Invalid short URL
          content: {}
          
    delete:
      tags:
        - short_links
      summary: Deletes data about specified short URL from the database
      operationId: Delete
      parameters:
      - name: shortID
        in: path
        description: Short URL the user wants to delete
        required: true
        schema:
          type: string
      security:
        - api_key: []
      responses:
        200:
          description: successful operation
          content: {}
        400:
          description: Invalid short URL supplied
          content: {}
      
  /{shortID}:
    get:
      tags:
        - short_links
      summary: Implements redirect to the long-URL that corresponds to specified short URL
      operationId: Redirect
      parameters:
      - name: shortID
        in: path
        description: the short URL user wants to convert to the long URL stored in th database
        required: true
        schema:
          type: string
      responses:
        308:
          description: Sends `Location` header with the long URL that corresponds to specified short URL
          headers:
            Location:
              description: New location, long URL.
              schema:
                type: string
        400:
          description: Invalid short URL supplied
          content: {}
  

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        password:
          type: string
          
    Link:
      type: object
      properties:
        shortID:
          type: string
        longURL:
          type: string
        description:
          type: string
        statistics:
          type: integer
          format: int64
          
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header 